+++ src/Resources/Resources/ChangeLog.md
@@ -21,6 +21,7 @@
 ## Upcoming Release
 * Added a notice to template deployment what-if output pointing users to Deployment Stacks What-If, which is now generally available and removes noise from results.
 * Renamed `DenySettingsApplyToChildScope` to `DenySettingsApplyToChildScopes` for deployment stack WhatIfResult cmdlets.
+* Added `ResourcesWithoutDeleteSupport` to deployment stack WhatIfResult cmdlets.
 
 ## Version 10.1.0
 * Added deployment stack WhatIfResult cmdlets for resource group, subscription, and management group scopes.